backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
python-multipart==0.0.6
requests==2.31.0
faker==22.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4